# 3d-backend/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"TableEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"tableName": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"columns": {
						"items": {
							"$ref": "#/components/schemas/ColumnEntity"
						},
						"type": "array"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/RowEntity"
						},
						"type": "array"
					},
					"cells": {
						"items": {
							"$ref": "#/components/schemas/CellEntity"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"tableName",
					"userId",
					"user",
					"columns",
					"rows",
					"cells",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CellEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"cellData": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"tableId": {
						"type": "number",
						"format": "double"
					},
					"table": {
						"$ref": "#/components/schemas/TableEntity"
					},
					"columnId": {
						"type": "number",
						"format": "double"
					},
					"column": {
						"$ref": "#/components/schemas/ColumnEntity"
					},
					"rowId": {
						"type": "number",
						"format": "double"
					},
					"row": {
						"$ref": "#/components/schemas/RowEntity"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"cellData",
					"type",
					"tableId",
					"table",
					"columnId",
					"column",
					"rowId",
					"row",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ColumnEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"columnName": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"table": {
						"$ref": "#/components/schemas/TableEntity"
					},
					"cells": {
						"items": {
							"$ref": "#/components/schemas/CellEntity"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"columnName",
					"type",
					"table",
					"cells",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RowEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"table": {
						"$ref": "#/components/schemas/TableEntity"
					},
					"cells": {
						"items": {
							"$ref": "#/components/schemas/CellEntity"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"table",
					"cells",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"table": {
						"items": {
							"$ref": "#/components/schemas/TableEntity"
						},
						"type": "array"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/RowEntity"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email",
					"table",
					"rows",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITablePayload": {
				"properties": {
					"tableName": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"tableName",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ColumnEntity_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"columnName": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"table": {
						"$ref": "#/components/schemas/TableEntity"
					},
					"cells": {
						"items": {
							"$ref": "#/components/schemas/CellEntity"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"Record_string.any_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ICellPayload": {
				"properties": {
					"cellData": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"tableId": {
						"type": "number",
						"format": "double"
					},
					"columnId": {
						"type": "number",
						"format": "double"
					},
					"rowId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"cellData",
					"type",
					"tableId",
					"columnId",
					"rowId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserPayload": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "3d-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/tables/{id}": {
			"get": {
				"operationId": "GetOneTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/TableEntity"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Table"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/TableEntity"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Table"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITablePayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Table"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/columns": {
			"post": {
				"operationId": "CreateColumn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ColumnEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Column"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"type": {
										"type": "string"
									},
									"tableId": {
										"type": "number",
										"format": "double"
									},
									"columnName": {
										"type": "string"
									}
								},
								"required": [
									"type",
									"tableId",
									"columnName"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllColumns",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ColumnEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Column"
				],
				"security": [],
				"parameters": []
			}
		},
		"/columns/{id}": {
			"get": {
				"operationId": "GetOneColumn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ColumnEntity"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Column"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateColumn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ColumnEntity"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Column"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ColumnEntity_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteColumn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Column"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/rows/{tableId}": {
			"post": {
				"operationId": "CreateRow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RowEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Row"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tableId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Record_string.any_"
							}
						}
					}
				}
			}
		},
		"/rows/{id}": {
			"get": {
				"operationId": "GetOneRow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/RowEntity"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Row"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateRow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/RowEntity"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Row"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Record_string.any_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Row"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/cells/bulk": {
			"post": {
				"operationId": "CreateBulkCells",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CellEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Cell"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/ICellPayload"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/cells": {
			"post": {
				"operationId": "CreateCell",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CellEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Cell"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICellPayload"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllCells",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CellEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Cell"
				],
				"security": [],
				"parameters": []
			}
		},
		"/cells/{id}": {
			"get": {
				"operationId": "GetOneCell",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CellEntity"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cell"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCell",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CellEntity"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cell"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICellPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCell",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Cell"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/cells/{tableId}/{rowId}": {
			"get": {
				"operationId": "GetCellsByRowId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CellEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Cell"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tableId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "rowId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}